test(autocomplete/tests): refactor test code for readability and add code documentation.
